+++ src/app/myworkflows/my-workflow/my-workflow.component.spec.ts
@@ -44,7 +44,6 @@ import {
   WorkflowStubService
 } from '../../test/service-stubs';
 import { RegisterWorkflowModalService } from '../../workflow/register-workflow-modal/register-workflow-modal.service';
-import { MyWorkflowsService } from '../myworkflows.service';
 import { MyWorkflowComponent } from './my-workflow.component';
 
 describe('MyWorkflowsComponent', () => {
@@ -76,7 +75,6 @@ describe('MyWorkflowsComponent', () => {
         { provide: AccountsService, useClass: AccountsStubService },
         { provide: WorkflowsService, useClass: WorkflowsStubService },
         { provide: UrlResolverService, useClass: UrlResolverStubService },
-        MyWorkflowsService,
         {
           provide: MatDialogRef,
           useValue: {
