+++ src/app/myworkflows/my-workflow/my-workflow.component.ts
@@ -67,7 +67,7 @@ import { MyWorkflowsService } from '../myworkflows.service';
   selector: 'app-my-workflow',
   templateUrl: './my-workflow.component.html',
   styleUrls: ['../../shared/styles/my-entry.component.scss'],
-  providers: [MyWorkflowsService, ProviderService, DockstoreService]
+  providers: [ProviderService, DockstoreService]
 })
 export class MyWorkflowComponent extends MyEntry implements OnInit {
   workflow: Service | BioWorkflow;
@@ -75,6 +75,7 @@ export class MyWorkflowComponent extends MyEntry implements OnInit {
   entryType: EntryType;
   entryType$: Observable<EntryType>;
   myEntryPageTitle$: Observable<string>;
+  workflow$: Observable<Service | BioWorkflow>;
   EntryType = EntryType;
   sharedWorkflows: Array<Workflow>;
   readonly pageName = '/my-workflows';
@@ -83,10 +84,8 @@ export class MyWorkflowComponent extends MyEntry implements OnInit {
   hasSourceControlToken$: Observable<boolean>;
   public gitHubAppInstallationLink$: Observable<string>;
   constructor(
-    private myworkflowService: MyWorkflowsService,
     protected configuration: Configuration,
     protected activatedRoute: ActivatedRoute,
-    private myBioWorkflowsService: MyBioWorkflowsService,
     private myServicesService: MyServicesService,
     private userQuery: UserQuery,
     private workflowService: WorkflowService,
@@ -102,7 +101,9 @@ export class MyWorkflowComponent extends MyEntry implements OnInit {
     private alertQuery: AlertQuery,
     private tokenService: TokenService,
     protected sessionService: SessionService,
-    protected sessionQuery: SessionQuery
+    protected sessionQuery: SessionQuery,
+    private myBioWorkflowsService: MyBioWorkflowsService,
+    private myWorkflowsService: MyWorkflowsService
   ) {
     super(accountsService, authService, configuration, tokenQuery, urlResolverService, sessionQuery, sessionService, activatedRoute);
     this.entryType = this.sessionQuery.getSnapshot().entryType;
@@ -110,6 +111,7 @@ export class MyWorkflowComponent extends MyEntry implements OnInit {
   }
 
   ngOnInit() {
+    this.myWorkflowsService.clearPartialState();
     this.gitHubAppInstallationLink$ = this.sessionQuery.gitHubAppInstallationLink$;
     this.tokenQuery.gitHubToken$
       .pipe(takeUntil(this.ngUnsubscribe))
@@ -120,6 +122,7 @@ export class MyWorkflowComponent extends MyEntry implements OnInit {
      * This handles selecting of a workflow based on changing URL. It also handles when the router changes url
      * due to when the user clicks the 'view checker workflow' or 'view parent entry' buttons.
      */
+    this.workflow$ = this.workflowQuery.selectActive();
     this.router.events
       .pipe(
         filter(event => event instanceof NavigationEnd),
@@ -136,7 +139,6 @@ export class MyWorkflowComponent extends MyEntry implements OnInit {
       });
     this.hasSourceControlToken$ = this.tokenQuery.hasSourceControlToken$;
     this.commonMyEntriesOnInit();
-    this.myworkflowService.clearPartialState();
 
     // Updates selected workflow from service and selects in sidebar
     this.workflowQuery.workflow$.pipe(takeUntil(this.ngUnsubscribe)).subscribe(workflow => (this.workflow = workflow));
@@ -149,11 +151,11 @@ export class MyWorkflowComponent extends MyEntry implements OnInit {
         ([workflows, sharedWorkflows]) => {
           if (workflows && sharedWorkflows) {
             this.workflows = workflows;
-            const sortedWorkflows = this.myworkflowService.sortGroupEntries(workflows, this.user.username, EntryType.BioWorkflow);
+            const sortedWorkflows = this.myWorkflowsService.sortGroupEntries(workflows, this.user.username, EntryType.BioWorkflow);
             this.setGroupEntriesObject(sortedWorkflows);
 
             this.sharedWorkflows = sharedWorkflows;
-            const sortedSharedWorkflows = this.myworkflowService.sortGroupEntries(
+            const sortedSharedWorkflows = this.myWorkflowsService.sortGroupEntries(
               sharedWorkflows,
               this.user.username,
               EntryType.BioWorkflow
@@ -186,11 +188,13 @@ export class MyWorkflowComponent extends MyEntry implements OnInit {
   }
 
   private getMyEntries() {
-    if (this.entryType === EntryType.BioWorkflow) {
-      this.myworkflowService.getMyEntries();
-    } else {
-      this.myServicesService.getMyEntries();
-    }
+    combineLatest(this.userQuery.user$, this.sessionQuery.entryType$)
+      .pipe(takeUntil(this.ngUnsubscribe))
+      .subscribe(([user, entryType]) => {
+        if (user && entryType) {
+          this.myWorkflowsService.getMyEntries(user.id, entryType);
+        }
+      });
   }
 
   /**
@@ -209,53 +213,15 @@ export class MyWorkflowComponent extends MyEntry implements OnInit {
   }
 
   protected convertOldNamespaceObjectToOrgEntriesObject(nsWorkflows: Array<any>): Array<OrgWorkflowObject> {
-    const groupEntriesObject: Array<OrgWorkflowObject> = [];
-    for (let i = 0; i < nsWorkflows.length; i++) {
-      const orgWorkflowObject: OrgWorkflowObject = {
-        sourceControl: '',
-        organization: '',
-        published: [],
-        unpublished: []
-      };
-      const nsWorkflow: Array<Workflow> = nsWorkflows[i].entries;
-      orgWorkflowObject.sourceControl = nsWorkflows[i].sourceControl;
-      orgWorkflowObject.organization = nsWorkflows[i].organization;
-      orgWorkflowObject.published = nsWorkflow.filter((workflow: Workflow) => {
-        return workflow.is_published;
-      });
-      orgWorkflowObject.unpublished = nsWorkflow.filter((workflow: Workflow) => {
-        return !workflow.is_published;
-      });
-      groupEntriesObject.push(orgWorkflowObject);
-    }
-    return groupEntriesObject;
+    return this.myWorkflowsService.convertOldNamespaceObjectToOrgEntriesObject(nsWorkflows);
   }
 
   protected getFirstPublishedEntry(orgWorkflows: Array<OrgWorkflowObject>): Workflow {
-    for (let i = 0; i < orgWorkflows.length; i++) {
-      const foundWorkflow = orgWorkflows[i]['entries'].find((workflow: Workflow) => {
-        return workflow.is_published === true;
-      });
-      if (foundWorkflow) {
-        return foundWorkflow;
-      }
-    }
-    return null;
+    return this.myWorkflowsService.getFirstPublishedEntry(orgWorkflows);
   }
 
   protected findEntryFromPath(path: string, orgWorkflows: Array<OrgWorkflowObject>): ExtendedWorkflow {
-    let matchingWorkflow: ExtendedWorkflow;
-    for (let i = 0; i < orgWorkflows.length; i++) {
-      matchingWorkflow = orgWorkflows[i].published.find((workflow: ExtendedWorkflow) => workflow.full_workflow_path === path);
-      if (matchingWorkflow) {
-        return matchingWorkflow;
-      }
-      matchingWorkflow = orgWorkflows[i].unpublished.find((workflow: ExtendedWorkflow) => workflow.full_workflow_path === path);
-      if (matchingWorkflow) {
-        return matchingWorkflow;
-      }
-    }
-    return null;
+    return this.myWorkflowsService.findEntryFromPath(path, orgWorkflows);
   }
 
   /**
@@ -290,9 +256,7 @@ export class MyWorkflowComponent extends MyEntry implements OnInit {
     this.registerWorkflowModalService.setWorkflowRepository(gitURL);
   }
 
-  showRegisterEntryModal(): void {
-    this.dialog.open(RegisterWorkflowModalComponent, { width: '600px' });
-  }
+  showRegisterEntryModal(): void {}
 
   refreshAllEntries(): void {
     this.refreshService.refreshAllWorkflows(this.user.id);
