+++ src/app/myworkflows/myworkflows.service.ts
@@ -14,17 +14,20 @@
  *    limitations under the License.
  */
 
-import { HttpErrorResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { MatSnackBar } from '@angular/material';
+import { MatDialog } from '@angular/material';
+import { transaction } from '@datorama/akita';
 import { AlertService } from 'app/shared/alert/state/alert.service';
-import { SessionQuery } from 'app/shared/session/session.query';
+import { EntryType } from 'app/shared/enum/entry-type';
+import { ExtendedWorkflow } from 'app/shared/models/ExtendedWorkflow';
+import { MyEntriesService } from 'app/shared/myentries.service';
 import { WorkflowService } from 'app/shared/state/workflow.service';
-import { UsersService, WorkflowsService } from 'app/shared/swagger';
+import { UsersService, Workflow, WorkflowsService } from 'app/shared/swagger';
 import { UserQuery } from 'app/shared/user/user.query';
-import { combineLatest, forkJoin, of as observableOf } from 'rxjs';
-import { catchError, finalize, takeUntil } from 'rxjs/operators';
-import { MyEntriesService } from './../shared/myentries.service';
+import { RegisterWorkflowModalComponent } from 'app/workflow/register-workflow-modal/register-workflow-modal.component';
+import { MyBioWorkflowsService } from './my-bio-workflows.service';
+import { MyServicesService } from './my-services.service';
+import { OrgWorkflowObject } from './my-workflow/my-workflow.component';
 
 @Injectable({
   providedIn: 'root'
@@ -32,21 +35,83 @@ import { MyEntriesService } from './../shared/myentries.service';
 export class MyWorkflowsService extends MyEntriesService {
   constructor(
     protected userQuery: UserQuery,
-    private sessionQuery: SessionQuery,
     protected alertService: AlertService,
     protected usersService: UsersService,
     protected workflowService: WorkflowService,
     protected workflowsService: WorkflowsService,
-    private matSnackBar: MatSnackBar
+    private myBioWorkflowsService: MyBioWorkflowsService,
+    private myServicesService: MyServicesService,
+    public matDialog: MatDialog
   ) {
     super();
+  }
+
+  getMyEntries(userId: number, entryType: EntryType) {
+    if (entryType === EntryType.BioWorkflow) {
+      this.myBioWorkflowsService.getMyBioWorkflows(userId);
+    } else {
+      this.myServicesService.getMyServices(userId);
+    }
   }
 
   getGroupIndex(groupEntries: any[], group: string): number {
     return groupEntries.findIndex(orgWorkflow => orgWorkflow.sourceControl + '/' + orgWorkflow.organization === group);
   }
 
+  public findEntryFromPath(path: string, orgWorkflows: Array<OrgWorkflowObject>): ExtendedWorkflow {
+    let matchingWorkflow: ExtendedWorkflow;
+    for (let i = 0; i < orgWorkflows.length; i++) {
+      matchingWorkflow = orgWorkflows[i].published.find((workflow: ExtendedWorkflow) => workflow.full_workflow_path === path);
+      if (matchingWorkflow) {
+        return matchingWorkflow;
+      }
+      matchingWorkflow = orgWorkflows[i].unpublished.find((workflow: ExtendedWorkflow) => workflow.full_workflow_path === path);
+      if (matchingWorkflow) {
+        return matchingWorkflow;
+      }
+    }
+    return null;
+  }
+
+  public getFirstPublishedEntry(orgWorkflows: Array<OrgWorkflowObject>): Workflow {
+    for (let i = 0; i < orgWorkflows.length; i++) {
+      const foundWorkflow = orgWorkflows[i]['entries'].find((workflow: Workflow) => {
+        return workflow.is_published === true;
+      });
+      if (foundWorkflow) {
+        return foundWorkflow;
+      }
+    }
+    return null;
+  }
+
+  public convertOldNamespaceObjectToOrgEntriesObject(nsWorkflows: Array<any>): Array<OrgWorkflowObject> {
+    const groupEntriesObject: Array<OrgWorkflowObject> = [];
+    for (let i = 0; i < nsWorkflows.length; i++) {
+      const orgWorkflowObject: OrgWorkflowObject = {
+        sourceControl: '',
+        organization: '',
+        published: [],
+        unpublished: []
+      };
+      const nsWorkflow: Array<Workflow> = nsWorkflows[i].entries;
+      orgWorkflowObject.sourceControl = nsWorkflows[i].sourceControl;
+      orgWorkflowObject.organization = nsWorkflows[i].organization;
+      orgWorkflowObject.published = nsWorkflow.filter((workflow: Workflow) => {
+        return workflow.is_published;
+      });
+      orgWorkflowObject.unpublished = nsWorkflow.filter((workflow: Workflow) => {
+        return !workflow.is_published;
+      });
+      groupEntriesObject.push(orgWorkflowObject);
+    }
+    return groupEntriesObject;
+  }
+
+  @transaction()
   clearPartialState(): void {
+    this.workflowService.clearActive();
+    this.workflowService.clearVersion();
     this.workflowService.setWorkflow(null);
     this.workflowService.setWorkflows(null);
     this.workflowService.setSharedWorkflows(null);
@@ -66,52 +131,7 @@ export class MyWorkflowsService extends MyEntriesService {
     return null;
   }
 
-  // Retrieve all of the workflows for the user and update the workflow service
-  // TODO: Fix this. What should happen is:
-  // If none of the two calls error, there should be a simple snackBar displayed
-  // If one of the two calls error, there should be a detailed card displayed
-  // If two of the calls error, there should be a weird combined detailed card displayed
-  // Any errors should still return an empty array for that set of workflows
-  getMyEntries(): void {
-    combineLatest(this.userQuery.user$, this.sessionQuery.entryType$)
-      .pipe(takeUntil(this.ngUnsubscribe))
-      .subscribe(([user, entryType]) => {
-        if (user && entryType) {
-          this.alertService.start('Fetching ' + entryType + 's');
-          this.getMyBioWorkflows(user.id);
-        }
-      });
+  registerEntry() {
+    this.matDialog.open(RegisterWorkflowModalComponent, { width: '600px' });
   }
-
-  getMyBioWorkflows(id: number): void {
-    forkJoin(
-      this.usersService.userWorkflows(id).pipe(
-        catchError((error: HttpErrorResponse) => {
-          this.alertService.detailedSnackBarError(error);
-          return observableOf([]);
-        })
-      ),
-      this.workflowsService.sharedWorkflows().pipe(
-        catchError((error: HttpErrorResponse) => {
-          this.alertService.detailedSnackBarError(error);
-          return observableOf([]);
-        })
-      )
-    )
-      .pipe(
-        finalize(() => this.alertService.simpleSuccess()),
-        takeUntil(this.ngUnsubscribe)
-      )
-      .subscribe(
-        ([workflows, sharedWorkflows]) => {
-          this.workflowService.setWorkflows(workflows);
-          this.workflowService.setSharedWorkflows(sharedWorkflows);
-        },
-        error => {
-          console.error('This should be impossible because both errors are caught already');
-        }
-      );
-  }
-
-  registerEntry(): void {}
 }
